requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-environ==0.10.0
Flask==2.2.3
Flask-RESTful==0.3.9
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
mpmath==1.3.0
numpy==1.23.5
oauthlib==3.2.2
Pillow==9.5.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyotp==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
sympy==1.11.1
urllib3==1.26.15
vk-api==11.9.9
Werkzeug==2.2.3
